Refactor TextFilter and add additional edge cases

+++ cogs/TextFilter.py
@@ -34,27 +34,36 @@ class TextFilterCog(commands.Cog):
 
     @commands.command(name="filter", aliases=["filter_word"])
     #@commands.check(KoalaBot.is_admin)
-    async def filter_new_word(self, ctx, word, filter_type="banned"):
+    async def filter_new_word(self, ctx, word, filter_type="banned", too_many_arguments=None):
         """
         Adds new word to the filtered text list
         :param ctx: The discord context
         :param word: The first argument and word to be filtered
+        :param filter_type: The filter type (banned or risky)
         :return:
         """
-        self.tf_database_manager.new_filtered_text(ctx.guild.id, word, filter_type)
-        await ctx.channel.send("*"+word+"* has been filtered.")
+        error = "Something has gone wrong, your word may already be filtered or you have entered the command incorrectly. You try again with: `k!filter [filtered_text] [[risky] or [banned]]`"
+        if too_many_arguments == None and typeExists(filter_type):
+            await filterWord(self, ctx, word, filter_type)
+            await ctx.channel.send("*" + word + "* has been filtered.")
+            return
+        raise Exception(error)
 
     @commands.command(name="unfilter", aliases=["unfilter_word"])
     #@commands.check(KoalaBot.is_admin)
-    async def unfilter_word(self, ctx, word):
+    async def unfilter_word(self, ctx, word, too_many_arguments=None):
         """
         Removes existing words from filter list
         :param ctx: The discord context
         :param word: The first argument and word to be filtered
         :return:
         """
-        self.tf_database_manager.unfilter_text(ctx.guild.id, word)
-        await ctx.channel.send("*"+word+"* has been unfiltered.")
+        error = "Too many arguments, please try again using the following arguments: `k!unfilter [filtered_word]`"
+        if too_many_arguments == None:
+            await unfilterWord(self, ctx, word)
+            await ctx.channel.send("*"+word+"* has been unfiltered.")
+            return
+        raise Exception(error)
 
     @commands.command(name="checkFilteredWords", aliases=["check_filtered_words"])
     #@commands.check(KoalaBot.is_admin)
@@ -68,10 +77,12 @@ class TextFilterCog(commands.Cog):
         embed.colour = KOALA_GREEN
         embed.set_footer(text=f"Guild ID: {ctx.guild.id}")
         embed.title = "Koala Moderation - Filtered Words"
-        words = ""
-        for word in self.tf_database_manager.get_filtered_text_for_guild(ctx.guild.id):
-            words+=word+"\n"
-        embed.add_field(name="Banned Words", value=words)
+        all_words, all_types = "", ""
+        for word, filter_type in self.tf_database_manager.get_filtered_text_for_guild(ctx.guild.id):
+            all_words+=word+"\n"
+            all_types+=filter_type+"\n"
+        embed.add_field(name="Banned Words", value=all_words)
+        embed.add_field(name="Filter Types", value=all_types)
 
         await ctx.channel.send(embed=embed)
 
@@ -150,10 +161,16 @@ class TextFilterCog(commands.Cog):
         """
         if (message.guild is not None): # and not KoalaBot.is_admin and not KoalaBot.is_owner):
             censor_list = self.tf_database_manager.get_filtered_text_for_guild(message.guild.id)
-            if (any(map(message.content.__contains__, censor_list))):
-                await message.author.send("Watch your language! Your message: '*"+message.content+"*' in "+message.channel.mention+" has been deleted by KoalaBot.")
-                await sendToModerationChannels(message, self)
-                await message.delete()
+            for word,filter_type in censor_list:
+                if (word in message.content):
+                    if (filter_type == "risky"):
+                        await message.author.send("Watch your language! Your message: '*"+message.content+"*' in "+message.channel.mention+" contains a 'risky' word. This is a warning.")
+                        break
+                    elif (filter_type == "banned"):
+                        await message.author.send("Watch your language! Your message: '*"+message.content+"*' in "+message.channel.mention+" has been deleted by KoalaBot.")
+                        await sendToModerationChannels(message, self)
+                        await message.delete()
+                        break
 
 def setup(bot: KoalaBot) -> None:
     """
@@ -161,6 +178,15 @@ def setup(bot: KoalaBot) -> None:
     :param  bot: The client of the KoalaBot
     """
     bot.add_cog(TextFilterCog(bot))
+
+async def filterWord(self, ctx, word, filter_type):
+    self.tf_database_manager.new_filtered_text(ctx.guild.id, word, filter_type)
+
+async def unfilterWord(self, ctx, word):
+    self.tf_database_manager.unfilter_text(ctx.guild.id, word)
+
+def typeExists(filter_type):
+    return filter_type == "risky" or filter_type == "banned"
 
 def isModerationChannelAvailable(guild_id, self):
     channels = self.tf_database_manager.get_mod_channel(guild_id)
@@ -181,7 +207,11 @@ def buildModerationEmbed(message):
     embed.add_field(name="User",value=message.author.mention)
     embed.add_field(name="Channel",value=message.channel.mention)
     embed.add_field(name="Message",value=message.content)
+    embed.add_field(name="Message",value=message.created_at)
     return embed
+
+def doesWordExist(self, ft_id):
+    return len(self.database_manager.db_execute_select(f"SELECT * FROM TextFilter WHERE filtered_text_id = (\"{ft_id}\");")) > 0
 
 class TextFilterDBManager:
     """
@@ -230,19 +260,22 @@ class TextFilterDBManager:
         """
         self.database_manager.db_execute_commit(
             f"INSERT INTO TextFilterModeration (channel_id, guild_id) VALUES (\"{channel_id}\", {guild_id});")
-    
+
     def new_filtered_text(self, guild_id, filtered_text, filter_type):
         """
         Adds new filtered word for a guild
         :param guild_id: Guild ID to retrieve filtered words from
         :param filtered_text: The new word to be filtered
+        :param filtered_type: The filter type (banned or risky)
         :return:
         """
         ft_id = str(guild_id) + filtered_text
-        self.database_manager.db_execute_commit(
-            f"INSERT INTO TextFilter (filtered_text_id, guild_id, filtered_text, filter_type) VALUES (\"{ft_id}\", {guild_id}, \"{filtered_text}\", \"{filter_type}\");")
-
-    
+        if not doesWordExist(self, ft_id):
+            self.database_manager.db_execute_commit(
+                f"INSERT INTO TextFilter (filtered_text_id, guild_id, filtered_text, filter_type) VALUES (\"{ft_id}\", {guild_id}, \"{filtered_text}\", \"{filter_type}\");")
+            return 
+        raise Exception("Filtered word already exists")
+            
     def unfilter_text(self, guild_id, filtered_text):
         """
         Adds new filtered word for a guild
@@ -251,8 +284,11 @@ class TextFilterDBManager:
         :return:
         """
         ft_id = str(guild_id) + filtered_text
-        self.database_manager.db_execute_commit(
-            f"DELETE FROM TextFilter WHERE filtered_text_id = (\"{ft_id}\");")
+        if doesWordExist(self, ft_id):
+            self.database_manager.db_execute_commit(
+                f"DELETE FROM TextFilter WHERE filtered_text_id = (\"{ft_id}\");")
+            return
+        raise Exception("Filtered word does not exist")
 
     def get_filtered_text_for_guild(self, guild_id):
         """
@@ -263,7 +299,7 @@ class TextFilterDBManager:
         rows = self.database_manager.db_execute_select(f"SELECT * FROM TextFilter WHERE guild_id = {guild_id};")
         censor_list = []
         for row in rows:
-            censor_list.append(row[2])
+            censor_list.append((row[2], row[3]))
         return censor_list
 
     def get_mod_channel(self, guild_id):
